feat: use FileFullscreenModal as a central, drop-in componet
