+++ src/script/components/MessagesList/Message/ContentMessage/asset/MultipartAssets/common/FileFullscreenModal/FileHeader/FileHeader.styles.ts
