+++ src/script/util/getFileTypeFromExtension/fileType.ts
